requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SocketIO>=5.3.0
python-socketio>=5.10.0
python-engineio>=4.8.0
Werkzeug==3.1.3
Jinja2==3.1.4
python-dotenv==1.0.0
typing-extensions>=4.12.0
pydantic==2.9.2
pydantic-core==2.23.4
gunicorn==21.2.0
supabase>=2.4.0
requests>=2.31.0
together  # Added for LLM_calls.together_get_response.py

# Vector Database
chromadb>=0.5.0

# RAG Orchestration
python-docx
pdfplumber
dateparser==1.2.2
